requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
PyPDF2==3.0.1
python-docx==1.1.2
typing_extensions==4.13.0
Werkzeug==3.1.3
